refactor: NotificationExeptionType 하나로 통일
