+++ backend/order/src/main/java/org/samtuap/inong/common/exceptionType/CouponExceptionType.java
@@ -1,0 +1,22 @@
+package org.samtuap.inong.common.exceptionType;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+public enum CouponExceptionType  implements ExceptionType{
+    COUPON_NOT_FOUND(HttpStatus.NOT_FOUND, "존재하지 않는 쿠폰입니다.");
+
+    private final HttpStatus status;
+    private final String message;
+
+    @Override
+    public HttpStatus httpStatus() {
+        return status;
+    }
+
+    @Override
+    public String message() {
+        return message;
+    }
+}
